refactor: split provider-bridge lib.rs into focused modules [OPE-489]
Split monolithic lib.rs (421 LOC) into:
- service.rs: GooseProviderService implementation
- types.rs: ConfigKeySummary, ProviderSummary types
- tests.rs: all test modules
